requirements.txt
flask
mysql-connector-python
requests
datetime
os
importlib
bcrypt